requirements.txt
django-crispy-forms~=1.6.0
django-import-export>=0.5.1
django-reversion~=2.0.0
django-formtools
future==0.15.2
httplib2==0.9.2
six==1.10.0
django-pure-pagination==0.3.0
django-simple-captcha
mysqlclient
requests
xlrd
XlsxWriter
xlwt
